pyproject.toml
[project]
name = "mell"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleksandr Sai",email = "alexcdsay@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django-bootstrap-v5 (>=1.0.11,<2.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
